+++ src/KorpiEngine.Runtime/Core/EntityModel/IDs/EntityID.cs
@@ -4,13 +4,13 @@ namespace KorpiEngine.Core.EntityModel.IDs;
 
 public static class EntityID
 {
-    private static ulong nextID;
+    private static int nextID;
     
     
-    public static ulong Generate()
+    public static int Generate()
     {
-        ulong id = Interlocked.Increment(ref nextID);
-        Debug.Assert(id != ulong.MaxValue, "EntityID overflow!");
+        int id = Interlocked.Increment(ref nextID);
+        Debug.Assert(id != int.MaxValue, "EntityID overflow!");
         return id;
     }
 }
